+++ src/types/sequence.rs
@@ -1,6 +1,6 @@
 // Copyright (c) 2017-present PyO3 Project and Contributors
 
-use crate::err::{self, PyDowncastError, PyErr, PyResult};
+use crate::err::{self, PyErr, PyResult};
 use crate::ffi::{self, Py_ssize_t};
 use crate::instance::PyNativeType;
 use crate::types::{PyAny, PyList, PyTuple};
@@ -11,6 +11,7 @@ use crate::{FromPyObject, PyTryFrom, ToBorrowedObject};
 #[repr(transparent)]
 pub struct PySequence(PyAny);
 pyobject_native_type_named!(PySequence);
+pyobject_native_type_info!(PySequence, ffi::PySequence_Type, Some("builtins"), #checkfunction=ffi::PySequence_Check);
 pyobject_native_type_extract!(PySequence);
 
 impl PySequence {
@@ -305,35 +306,12 @@ where
     Ok(v)
 }
 
-impl<'v> PyTryFrom<'v> for PySequence {
-    fn try_from<V: Into<&'v PyAny>>(value: V) -> Result<&'v PySequence, PyDowncastError<'v>> {
-        let value = value.into();
-        unsafe {
-            if ffi::PySequence_Check(value.as_ptr()) != 0 {
-                Ok(<PySequence as PyTryFrom>::try_from_unchecked(value))
-            } else {
-                Err(PyDowncastError::new(value, "Sequence"))
-            }
-        }
-    }
-
-    fn try_from_exact<V: Into<&'v PyAny>>(value: V) -> Result<&'v PySequence, PyDowncastError<'v>> {
-        <PySequence as PyTryFrom>::try_from(value)
-    }
-
-    #[inline]
-    unsafe fn try_from_unchecked<V: Into<&'v PyAny>>(value: V) -> &'v PySequence {
-        let ptr = value.into() as *const _ as *const PySequence;
-        &*ptr
-    }
-}
-
 #[cfg(test)]
 mod test {
     use crate::types::PySequence;
     use crate::AsPyPointer;
     use crate::Python;
-    use crate::{PyObject, PyTryFrom, ToPyObject};
+    use crate::{Py, PyObject, PyTryFrom, ToPyObject};
 
     fn get_object() -> PyObject {
         // Convenience function for getting a single unique object
@@ -663,7 +641,7 @@ mod test {
     }
 
     #[test]
-    fn test_is_empty() {
+    fn test_as_ref() {
         let gil = Python::acquire_gil();
         let py = gil.python();
         let list = vec![1].to_object(py);
@@ -673,5 +651,14 @@ mod test {
         let empty_list = vec.to_object(py);
         let empty_seq = empty_list.cast_as::<PySequence>(py).unwrap();
         assert!(empty_seq.is_empty().unwrap());
+    }
+
+    #[test]
+    fn test_is_empty() {
+        let gil = Python::acquire_gil();
+        let py = gil.python();
+        let list = vec!["foo"].to_object(py);
+        let seq: Py<PySequence> = list.cast_as::<PySequence>(py).unwrap().into();
+        assert!(seq.as_ref(py).extract::<Vec<&str>>().unwrap() == ["foo"]);
     }
 }